requirements.txt
# Package use to serve an API
schematics==1.1.1
supercell

# Package use for NLP
spacy

# Other useful stuff
PyYAML
